perf(theme): move INDUSTRIAL ring off the React render path, drop central mousePos
The warning ring already moved via transform but still re-rendered
through React state every mousemove. It now follows the cursor via
ref-written translate3d like every other mouse theme.

With its last consumer migrated, the ForegroundLayer central
mousemove-to-setState machinery and the shared MousePos prop interface
are removed entirely - no theme re-renders on cursor movement anymore.
