+++ src/components/TerminalUI.tsx
@@ -4,7 +4,7 @@
  * 负责根据当前主题渲染对应的背景和前景效果
  * 主题效果组件位于 ./themes/ 目录
  */
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { useIsMobile } from "../hooks/useIsMobile";
 import { ThemePreset } from "../types";
 // 主题效果
@@ -70,30 +70,13 @@ export const BackgroundLayer: React.FC<{ theme?: ThemePreset }> = ({
 export const ForegroundLayer: React.FC<{ theme?: ThemePreset }> = ({
     theme = ThemePreset.ORIGIN,
 }) => {
-    const [mousePos, setMousePos] = useState({ x: 0, y: 0 });
     const isMobile = useIsMobile();
-
-    // 仅 INDUSTRIAL 仍通过 React state 消费鼠标坐标；
-    // 其余鼠标主题均已迁移到 ref 直写 transform，不再触发逐帧重渲染
-    const needsMousePos = !isMobile && theme === ThemePreset.INDUSTRIAL;
-
-    useEffect(() => {
-        if (!needsMousePos) return;
-
-        // 同步 setState：浏览器按帧节奏合并派发 mousemove（每帧至多一次），
-        // 经 rAF 转发可能多等一帧，表现为光标跟随延迟
-        const handleMouseMove = (e: MouseEvent) => {
-            setMousePos({ x: e.clientX, y: e.clientY });
-        };
-        window.addEventListener("mousemove", handleMouseMove);
-        return () => {
-            window.removeEventListener("mousemove", handleMouseMove);
-        };
-    }, [needsMousePos]);
 
     // 移动端不渲染鼠标交互层
     if (isMobile) return null;
 
+    // 所有鼠标跟随主题均在组件内部以 ref 直写 transform 跟踪光标，
+    // 不再经由 React state 逐帧重渲染
     switch (theme) {
         case ThemePreset.ORIGIN:
             return <OriginForeground />;
@@ -102,7 +85,7 @@ export const ForegroundLayer: React.FC<{ theme?: ThemePreset }> = ({
         case ThemePreset.ABYSSAL:
             return <AbyssalForeground />;
         case ThemePreset.INDUSTRIAL:
-            return <IndustrialForeground mousePos={mousePos} />;
+            return <IndustrialForeground />;
         case ThemePreset.AZURE:
             return <AzureForeground />;
         case ThemePreset.MIKU:
